hw_29/requirements.txt
pillow==11.1.0
pillow-heif==0.21.0